fix: centralize sandbox tool policy
