+++ packages/wordpress-plugin/src/generated-sandbox-datamachine-tool-policy.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * Generated sandbox Data Machine tool policy.
+ *
+ * Source: packages/sandbox-datamachine-tool-policy.json.
+ *
+ * @package WPCodebox
+ */
+
+return array(
+	'schema' => 'wp-codebox/sandbox-datamachine-tool-policy/v1',
+	'version' => 1,
+	'safe_tools' => array(
+		'datamachine/workspace-read',
+		'datamachine/workspace-ls',
+		'datamachine/workspace-grep',
+		'datamachine/workspace-write',
+		'datamachine/workspace-edit',
+		'datamachine/workspace-apply-patch',
+		'datamachine/workspace-git-status',
+		'datamachine/workspace-git-log',
+		'datamachine/workspace-git-diff',
+		'datamachine/list-github-issues',
+		'datamachine/get-github-issue',
+		'datamachine/list-github-pulls',
+		'datamachine/get-github-pull',
+		'datamachine/list-github-pull-files',
+		'datamachine/get-github-check-runs',
+		'datamachine/get-github-commit-statuses',
+		'datamachine/list-github-tree',
+		'datamachine/get-github-file',
+		'datamachine/list-github-repos',
+	),
+	'parent_only_tools' => array(
+		'datamachine/workspace-clone',
+		'datamachine/workspace-adopt',
+		'datamachine/workspace-remove',
+		'datamachine/workspace-delete',
+		'datamachine/workspace-git-pull',
+		'datamachine/workspace-git-add',
+		'datamachine/workspace-git-commit',
+		'datamachine/workspace-git-push',
+		'datamachine/workspace-git-rebase',
+		'datamachine/workspace-git-reset',
+		'datamachine/workspace-pr-rebase',
+		'datamachine/workspace-worktree-add',
+		'datamachine/workspace-worktree-finalize',
+		'datamachine/workspace-worktree-remove',
+		'datamachine/workspace-worktree-prune',
+		'datamachine/workspace-worktree-cleanup',
+		'datamachine/workspace-cleanup-apply',
+		'datamachine/create-github-issue',
+		'datamachine/update-github-issue',
+		'datamachine/create-github-pull-request',
+		'datamachine/comment-github-issue',
+		'datamachine/comment-github-pull-request',
+		'datamachine/upsert-github-pull-review-comment',
+		'datamachine/merge-github-pull-request',
+		'datamachine/cleanup-github-pull-request',
+		'datamachine/create-or-update-github-file',
+		'datamachine/create-code-task',
+		'datamachine/gitsync-bind',
+		'datamachine/gitsync-unbind',
+		'datamachine/gitsync-pull',
+		'datamachine/gitsync-submit',
+		'datamachine/gitsync-push',
+		'datamachine/gitsync-policy-update',
+	),
+);
